fixing Cluster IT

+++ server/src/main/java/org/apache/iotdb/db/engine/load/LoadTsFileManager.java
@@ -177,7 +177,9 @@ public class LoadTsFileManager {
     }
 
     private void clearDir(File dir) {
-      FileUtils.deleteDirectory(dir);
+      if (dir.exists()) {
+        FileUtils.deleteDirectory(dir);
+      }
       if (dir.mkdirs()) {
         logger.info(String.format("Load TsFile dir %s is created.", dir.getPath()));
       }
@@ -210,6 +212,9 @@ public class LoadTsFileManager {
     }
 
     private void writeDeletion(TsFileData deletionData) throws IOException {
+      if (isClosed) {
+        throw new IOException(String.format("%s TsFileWriterManager has been closed.", taskDir));
+      }
       for (Map.Entry<DataPartitionInfo, TsFileIOWriter> entry : dataPartition2Writer.entrySet()) {
         deletionData.writeToFileWriter(entry.getValue());
       }
@@ -247,6 +252,9 @@ public class LoadTsFileManager {
     }
 
     private void close() {
+      if (isClosed) {
+        return;
+      }
       if (dataPartition2Writer != null) {
         for (Map.Entry<DataPartitionInfo, TsFileIOWriter> entry : dataPartition2Writer.entrySet()) {
           try {
@@ -254,7 +262,7 @@ public class LoadTsFileManager {
             if (writer.canWrite()) {
               writer.close();
             }
-            if (!writer.getFile().delete()) {
+            if (writer.getFile().exists() && !writer.getFile().delete()) {
               logger.warn(String.format("Delete File %s error.", writer.getFile()));
             }
           } catch (IOException e) {
